Poised-CRUDO Testing Strategy : Admin Tests : Read with Parameters : Part 12 : New Test Case Added : Make Several Read Requests With Different Headers
In this task, we read the contents of
admin_read_requests_parameterized.txt as ${READ_REQUESTS}:

[[{'Host': 'glacial-earth-31542.herokuapp.com'}, 200], [{'Accept':
'application/json'}, 200], [{}, 200]]

Each inner list item, for example:

[{'Host': 'glacial-earth-31542.herokuapp.com'}, 200]

has a read request header combination:
{'Host': 'glacial-earth-31542.herokuapp.com'}

and expected response code: i.e. 200.

We make a read request for each read request header combination and
store the response code as the 3.rd item into the inner list item.
For example, suppose that we received 400 as a response code for the
above inner list item, then the updated inner list item will be:

[{'Host': 'glacial-earth-31542.herokuapp.com'}, 200, 400]

Imagine we make calls for all the inner list items, then
${READ_REQUESTS} would be updated for example:

[[{'Host': 'glacial-earth-31542.herokuapp.com'}, 200, 400], [{'Accept':
'application/json'}, 200, 200], [{}, 200, 200]]

The update logic takes place in
AdminUser.make_multiple_get_requests_with_different_headers() method.

After the update take place, the newly added test case checks if the
expected vs observed response codes are same for all the requests.

If so test passes, if not, the test case logs the mismatches as error
and fails.